+++ Source/Contracts/ContractTest.swift
@@ -6,8 +6,8 @@ final class СontractAddressTest: XCTestCase {
         // should resolve address correctly
         
         let stateInit = StateInit(
-            code: try Builder().storeUint(UInt32(1), bits: 8).endCell(),
-            data: try Builder().storeUint(UInt32(2), bits: 8).endCell()
+            code: try Builder().storeUint(UInt64(1), bits: 8).endCell(),
+            data: try Builder().storeUint(UInt64(2), bits: 8).endCell()
         )
         let addr = try OpaqueContract(workchain: 0, stateInit: stateInit).address()
         XCTAssertEqual(addr, try Address.parse(source: "EQCSY_vTjwGrlvTvkfwhinJ60T2oiwgGn3U7Tpw24kupIhHz"))
